+++ dash_core_components_base/__init__.py
@@ -45,6 +45,7 @@ async_resources = [
     'graph',
     'highlight',
     'markdown',
+    'slider',
     'upload'
 ]
