+++ src/components/RangeSlider.react.js
@@ -1,113 +1,19 @@
-import React, {Component} from 'react';
+import React, {Component, lazy, Suspense} from 'react';
 import PropTypes from 'prop-types';
-import {assoc, omit, pickBy} from 'ramda';
-import {Range, createSliderWithTooltip} from 'rc-slider';
-import computeSliderStyle from '../utils/computeSliderStyle';
+import rangeSlider from '../utils/LazyLoader/rangeSlider';
 
-import 'rc-slider/assets/index.css';
+const RealRangeSlider = lazy(rangeSlider);
 
 /**
  * A double slider with two handles.
  * Used for specifying a range of numerical values.
  */
 export default class RangeSlider extends Component {
-    constructor(props) {
-        super(props);
-        this.propsToState = this.propsToState.bind(this);
-        this.DashSlider = props.tooltip
-            ? createSliderWithTooltip(Range)
-            : Range;
-        this._computeStyle = computeSliderStyle();
-    }
-
-    propsToState(newProps) {
-        this.setState({value: newProps.value});
-    }
-
-    componentWillReceiveProps(newProps) {
-        if (newProps.tooltip !== this.props.tooltip) {
-            this.DashSlider = newProps.tooltip
-                ? createSliderWithTooltip(Range)
-                : Range;
-        }
-        this.propsToState(newProps);
-    }
-
-    componentWillMount() {
-        this.propsToState(this.props);
-    }
-
     render() {
-        const {
-            className,
-            id,
-            loading_state,
-            setProps,
-            tooltip,
-            updatemode,
-            vertical,
-            verticalHeight,
-        } = this.props;
-        const value = this.state.value;
-
-        let tipProps;
-        if (tooltip && tooltip.always_visible) {
-            /**
-             * clone `tooltip` but with renamed key `always_visible` -> `visible`
-             * the rc-tooltip API uses `visible`, but `always_visible is more semantic
-             * assigns the new (renamed) key to the old key and deletes the old key
-             */
-            tipProps = assoc('visible', tooltip.always_visible, tooltip);
-            delete tipProps.always_visible;
-        } else {
-            tipProps = tooltip;
-        }
-
-        const truncatedMarks =
-            this.props.marks &&
-            pickBy(
-                (k, mark) => mark >= this.props.min && mark <= this.props.max,
-                this.props.marks
-            );
-
         return (
-            <div
-                id={id}
-                data-dash-is-loading={
-                    (loading_state && loading_state.is_loading) || undefined
-                }
-                className={className}
-                style={this._computeStyle(vertical, verticalHeight, tooltip)}
-            >
-                <this.DashSlider
-                    onChange={value => {
-                        if (updatemode === 'drag') {
-                            setProps({value});
-                        } else {
-                            this.setState({value});
-                        }
-                    }}
-                    onAfterChange={value => {
-                        if (updatemode === 'mouseup') {
-                            setProps({value});
-                        }
-                    }}
-                    tipProps={tipProps}
-                    value={value}
-                    marks={truncatedMarks}
-                    {...omit(
-                        [
-                            'className',
-                            'value',
-                            'setProps',
-                            'marks',
-                            'updatemode',
-                            'verticalHeight',
-                        ],
-                        this.props
-                    )}
-                />
-            </div>
+            <Suspense fallback={null}>
+                <RealRangeSlider {...this.props} />
+            </Suspense>
         );
     }
 }
@@ -308,3 +214,6 @@ RangeSlider.defaultProps = {
     persistence_type: 'local',
     verticalHeight: 400,
 };
+
+export const propTypes = RangeSlider.propTypes;
+export const defaultProps = RangeSlider.defaultProps;
