+++ src/utils/LazyLoader/rangeSlider.js
@@ -1,0 +1,2 @@
+export default () => import(/* webpackChunkName: "slider" */ '../../fragments/RangeSlider.react');
+
