+++ src/utils/LazyLoader/slider.js
@@ -1,0 +1,2 @@
+export default () => import(/* webpackChunkName: "slider" */ '../../fragments/Slider.react');
+
